# openapi.yaml
openapi: 3.1.0
info:
  title: DeepResearch Agent API
  version: 1.5.0
  description: >
    HTTP wrapper for the DeepResearch ReAct workflow. Returns the final report,
    step trace, structured observations, deduplicated references, and built-in
    skill metadata.
servers:
  - url: https://buccaneer-recent-dropout.ngrok-free.dev
    description: Public tunnel to the local service
paths:
  /health:
    get:
      operationId: healthCheck
      summary: Health check
      responses:
        "200":
          description: Service is available
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /skills:
    get:
      operationId: listBuiltInSkills
      summary: List available built-in skills
      description: >
        Return the built-in skill catalog used by the ReAct agent, including
        category, required arguments, optional arguments, descriptions,
        current enabled state, and available skill profiles.
      responses:
        "200":
          description: Skill catalog
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_skills:
                    type: integer
                    description: Number of registered built-in skills
                  enabled_skills:
                    type: integer
                    description: Number of enabled built-in skills
                  categories:
                    type: array
                    description: Skill categories
                    items:
                      type: string
                  profiles:
                    type: array
                    description: Available skill profiles
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        description:
                          type: string
                        allowed_skills:
                          type: array
                          items:
                            type: string
                        allowed_count:
                          type: integer
                  skills:
                    type: array
                    description: Available built-in skills
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        description:
                          type: string
                        category:
                          type: string
                        required_args:
                          type: array
                          items:
                            type: string
                        optional_args:
                          type: array
                          items:
                            type: string
                        args_desc:
                          type: object
                          additionalProperties:
                            type: string
                        returns_sources:
                          type: boolean
                        enabled:
                          type: boolean
  /skills/route-preview:
    post:
      operationId: previewSkillRoute
      summary: Preview skill routing for a question
      description: >
        Preview the current RouteDecision for a question after applying enabled
        skills, the selected profile allowlist, and question routing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                question:
                  type: string
                  description: Question to preview routing for
                engine:
                  type: string
                  description: Engine preset. Use "deep", "fast", or empty for auto.
                  default: ""
                  enum: ["", "deep", "fast"]
                skill_profile:
                  type: string
                  description: Skill exposure profile used to preview visible skills.
                  default: react_default
      responses:
        "200":
          description: Route preview
          content:
            application/json:
              schema:
                type: object
                properties:
                  question:
                    type: string
                  question_type:
                    type: string
                  skill_profile:
                    type: string
                  allowed_skills:
                    type: array
                    items:
                      type: string
                  preferred_skills:
                    type: array
                    items:
                      type: string
                  discouraged_skills:
                    type: array
                    items:
                      type: string
                  starter:
                    type: string
                  reasons:
                    type: array
                    items:
                      type: string
                  signals:
                    type: array
                    items:
                      type: string
        "400":
          description: Invalid request
  /run:
    post:
      operationId: runDeepResearchAgent
      summary: Run DeepResearch Agent
      description: >
        Execute the ReAct research workflow for a question and return the final
        markdown report, step trace, structured observations, and references.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                question:
                  type: string
                  description: Research question
                  example: What were Tesla's core financial metrics in 2024?
                engine:
                  type: string
                  description: Engine preset. Use "deep", "fast", or empty for auto.
                  default: ""
                  enum: ["", "deep", "fast"]
                max_steps:
                  type: integer
                  description: Maximum number of ReAct steps
                  default: 8
                  minimum: 3
                  maximum: 15
                skill_profile:
                  type: string
                  description: Skill exposure profile used to filter available skills.
                  default: api_safe
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  answer:
                    type: string
                    description: Final markdown report
                  step_count:
                    type: integer
                    description: Actual number of executed steps
                  error:
                    type: string
                    nullable: true
                    description: Execution error, if any
                  references_md:
                    type: string
                    description: Markdown reference section
                  skill_profile:
                    type: string
                    description: Resolved skill profile used during execution
                  references:
                    type: array
                    description: Deduplicated references
                    items:
                      type: object
                      properties:
                        cite_id:
                          type: integer
                        url:
                          type: string
                        title:
                          type: string
                        snippet:
                          type: string
                  observations:
                    type: array
                    description: Structured observations collected during execution
                    items:
                      type: object
                      properties:
                        content:
                          type: string
                        tool:
                          type: string
                        args:
                          type: object
                          additionalProperties: true
                        cite_ids:
                          type: array
                          items:
                            type: integer
                        sources:
                          type: array
                          items:
                            type: object
                            properties:
                              url:
                                type: string
                              title:
                                type: string
                              snippet:
                                type: string
                  steps:
                    type: array
                    description: ReAct step trace
                    items:
                      type: object
                      properties:
                        thought:
                          type: string
                        tool:
                          type: string
                        args:
                          type: object
                          additionalProperties: true
                        observation:
                          type: string
                        cite_ids:
                          type: array
                          items:
                            type: integer
                        error_type:
                          type: string
                          nullable: true
                        sources:
                          type: array
                          items:
                            type: object
                            properties:
                              url:
                                type: string
                              title:
                                type: string
                              snippet:
                                type: string
        "400":
          description: Invalid request